added conditional compile options
