+++ libOTe/Dpf/DpfMult.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "libOTe/config.h"
+#if defined(ENABLE_REGULAR_DPF) || defined(ENABLE_SPARSE_DPF)
 
 #include "cryptoTools/Common/Defines.h"
 #include "coproto/Socket/Socket.h"
@@ -163,5 +165,6 @@ namespace osuCrypto
 	};
 
 }
-
 #undef SIMD8
+
+#endif // defined(ENABLE_REGULAR_DPF) || defined(ENABLE_SPARSE_DPF)
