+++ libOTe/Dpf/RegularDpf.h
@@ -1,5 +1,8 @@
 #pragma once
 
+#include "libOTe/config.h"
+#if defined(ENABLE_REGULAR_DPF)
+
 
 #include "cryptoTools/Common/Defines.h"
 #include "coproto/Socket/Socket.h"
@@ -756,3 +759,5 @@ namespace osuCrypto
 }
 
 #undef SIMD8
+
+#endif
