+++ libOTe/Triple/Foleage/FoleageTriple.cpp
@@ -1,8 +1,21 @@
+// © 2025 Peter Rindal.
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+// Code partially authored by:
+// Maxime Bombar, Dung Bui, Geoffroy Couteau, Alain Couvreur, Clément Ducros, and Sacha Servan - Schreiber
+
+#include "libOTe/config.h"
+#if defined(ENABLE_FOLEAGE)
+
 #include "FoleageTriple.h"
 #include "libOTe/Triple/Foleage/FoleageUtils.h"
 #include "libOTe/Triple/Foleage/fft/FoleageFft.h"
 #include "cryptoTools/Common/BitIterator.h"
-#include "libOTe/Dpf/TriDpf.h"
+#include "libOTe/Dpf/TernaryDpf.h"
 #include "libOTe/Base/BaseOT.h"
 namespace osuCrypto
 {
@@ -533,7 +546,7 @@ namespace osuCrypto
 		}
 		setTimePoint("transpose");
 
-		fft_recursive_uint32(fft, mLog3N, mN / 3);
+		foleageFftUint32(fft, mLog3N, mN / 3);
 		setTimePoint("product fft");
 		F4Multiply(mFftASquared, fft, fftRes, mN);
 		setTimePoint("product mult");
@@ -739,3 +752,4 @@ namespace osuCrypto
 	//}
 
 }
+#endif
