+++ libOTe/Triple/Foleage/fft/FoleageFft.h
@@ -1,4 +1,16 @@
 #pragma once
+// © 2025 Peter Rindal.
+// Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+// 
+// The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+// 
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+// Code partially authored by:
+// Maxime Bombar, Dung Bui, Geoffroy Couteau, Alain Couvreur, Clément Ducros, and Sacha Servan - Schreiber
+
+#include "libOTe/config.h"
+#if defined(ENABLE_FOLEAGE)
 
 #include <string.h>
 #include <stdint.h>
@@ -8,23 +20,20 @@
 
 namespace osuCrypto {
 
-	//typedef __int128 int128_t;
-	//typedef unsigned __int128 uint128_t;
-
 	// FFT for (up to) 32 polynomials over F4
-	void fft_recursive_uint64(
+	void foleageFftUint64(
 		span<uint64_t> coeffs,
 		const size_t num_vars,
 		const size_t num_coeffs);
 
 	// FFT for (up to) 16 polynomials over F4
-	void fft_recursive_uint32(
+	void foleageFftUint32(
 		span<uint32_t> coeffs,
 		const size_t num_vars,
 		const size_t num_coeffs);
 
 	// FFT for (up to) 8 polynomials over F4
-	void fft_recursive_uint16(
+	void foleageFftUint16(
 		span<uint16_t> coeffs,
 		const size_t num_vars,
 		const size_t num_coeffs);
@@ -37,3 +46,5 @@ namespace osuCrypto {
 
 
 }
+
+#endif
